introducing Column
No new concepts here, this new set of classes is just a renaming of TableField, NdFieldBase, and NdField. The overuse of the "Field" name element was somewhat confusing. Now the meaning of the "primitive" object pointed-to by Table (ColumnBase), is well separated from those objects which are intended to be named members under a Table-derived class deinition (NdFieldGroup and NdField)