Reverting to copy semantics for Fields
The MultiField has a tuple containing the values of the fields it contains, and although these are accessible by std::get, it is more meaningful to define aliases as references.

This has a significant drawback in that the copy ctor for a tuple is essentially a call to a function with multiple arguments, which the compiler is free to execute in any order. This is a problem when the copy ctors of the elements of the tuple have side effects - which of course is the case with Fields. Hence, the Fields might not necessarily get added to the Row in the same order in which they appear in the tuple. This is not acceptable, although the MultiField was not subject to this in practise because its Field-typed members were always passed in initially with a null Row pointer to be (added to the Row in tuple-order in the MultiField ctor), and the FieldBase did not cause side effects in this state. Requiring the FieldBase to respect this "before placement in Row" state is a bit untidy, and more significantly, the MultiField is only able to have FieldBase-derived objects as its members, precluding the aggregation of on MultiField within another to create more complex data structures (the outer-product basis function of a pair of FermionONV-BosonONV product being the motivation for this in the present development)

The new CompositeField solves the problem by inverting the relationship between tuple and member. This means there needs to be a class definition for each CompositeField with copy ctor and copy assignment operator. This is not a significant negative, and in fact MultiField specializations were inherited into classes such as FrmBosOnvField anyway. The members store the values, and the tuple stores references to these, not the other way around as was the case with the MultiField. This avoids the copy order problem and allows of arbitrary-depth aggregation of CompositeFields.