implementing and testing heisenberg exchange conn_foreach
also simplifying the interface for setting and adding operators to connections. These methods are called on the creation and annihilation strings separately, and those methods which acted upon the entire connection have been removed. This leads to an improvement in readability. The ability to compare a pair of indices before adding them to a fermion op product is also added in the new set_in_order method